Refactored webpack config
