+++ packages/chain-events/.eslintrc.json
@@ -15,7 +15,7 @@
     "import/resolver": {
       "node": {},
       "webpack": {
-        "config": "webpack/webpack.common.js"
+        "config": "webpack/webpack.base.config.js"
       }
     }
   },
