+++ packages/commonwealth/.eslintrc.json
@@ -20,7 +20,7 @@
     "import/resolver": {
       "node": {},
       "webpack": {
-        "config": "webpack/webpack.common.js"
+        "config": "webpack/webpack.base.config.js"
       }
     }
   },
