+++ packages/commonwealth/stats.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
 # Generates and visualizes webpack module sizes
-webpack --json --mode=development --config webpack/webpack.config.prod.js > stats.json
+webpack --json --mode=development --config webpack/webpack.prod.config.js > stats.json
 yarn webpack-bundle-analyzer stats.json
